Plugging sparse_utils to Facility Location code

+++ cpp/submod/wr_FacilityLocation.cpp
@@ -12,7 +12,7 @@ void cl_FacilityLocation(py::module &m)
     py::class_<FacilityLocation>(m,"FacilityLocation")
         //constructor(no_of_elem_in_ground, mode, sim_matrix or cluster, num_neigh, partial, ground_subset )
         .def(py::init<ll, std::string, std::vector<std::vector<float>>, ll, bool, std::set<ll> >()) //dense matrix
-        .def(py::init<ll, std::string, std::vector<float>, std::vector<float>, std::vector<float>, ll, bool, std::set<ll>>()) //sparse matrix
+        .def(py::init<ll, std::string, std::vector<float>, std::vector<ll>, std::vector<ll>, ll, bool, std::set<ll>>()) //sparse matrix
         .def(py::init<ll, std::string, std::vector<std::set<ll>>, ll, bool, std::set<ll>>()) //cluster
         .def("evaluate", &FacilityLocation::evaluate)
         .def("evaluateSequential", &FacilityLocation::evaluateSequential)
